+++ src/main/shared/events/index.ts
@@ -95,6 +95,8 @@ export const Events = {
         onServer: 'webview:on:server',
         onEmit: 'webview:emit:on',
         onKeypress: 'webview:emit:keypress',
+        emitClientRpc: 'webview:emit:client:rpc',
+        emitServerRpc: 'webview:emit:server:rpc',
         emitServer: 'webview:emit:server',
         emitClient: 'webview:emit:client',
         emitReady: 'webview:emit:page:ready',